- Added ImmediateGeometryDX11, which provides an easy to use geometry class for rendering dynamic shapes. - Added ImmediateRenderer project, which demonstrates how to use the new ImmediateGeometryDX11 class. - Refactored a couple things in the PipelineExecutorDX11 class from the GeometryDX11 class. - Added a simple Draw method to the PipelineManagerDX11 class to support the new geometry class. - Added an extension system with separate project for the Kinect libraries.  This will eventually lead to separate solutions for the extension based samples. - Removed un-needed <memory.h> includes, and added PCH includes where they were needed. - Improved and tested the ShaderFactoryDX11 method for loading compiled shaders.

+++ trunk/Hieroglyph3/Source/Objects/Camera.cpp
@@ -24,7 +24,7 @@ Camera::Camera()
 	m_fNear = 0.1f;
 	m_fFar = 100.0f;
 	m_fAspect = 1280.0f / 800.0f;
-	m_fFov = GLYPH_PI / 4.0f;
+	m_fFov = static_cast<float>( GLYPH_PI ) / 4.0f;
 
     m_ProjMatrix.MakeIdentity();
 }
